getLastMesure dans classe capteur

+++ src/main/java/com/ptut/penibilite/daos/CapteurRepository.java
@@ -23,5 +23,5 @@ public interface CapteurRepository extends JpaRepository<Capteur, Integer>{
      * @return la derniere mesure relevée
      */
     @Query(value = "SELECT M.valeur from Mesure M where M.capteur_id = :id and M.date = (Select Max(Me.date) FROM Mesure Me WHERE Me.date <= :date )", nativeQuery = true)
-    int getLastMesure(int id, LocalDateTime date);
+    Float getLastMesure(int id, LocalDateTime date);
 }
